requirements.txt
feedparser==6.0.12
requests==2.32.5
python-dotenv==1.2.1